tenant_id data type change to long

+++ backend/ModernWMS.WMS/Entities/Models/Customer/CustomerEntity.cs
@@ -63,7 +63,7 @@ namespace ModernWMS.WMS.Entities.Models
         /// <summary>
         /// the tenant id
         /// </summary>
-        public byte tenant_id { get; set; } = 1;
+        public long tenant_id { get; set; } = 1;
         #endregion
     }
 }
