+++ backend/ModernWMS.WMS/Entities/Models/Sku/CategoryEntity.cs
@@ -50,7 +50,7 @@ namespace ModernWMS.WMS.Entities.Models
         /// <summary>
         /// tenant_id
         /// </summary>
-        public byte tenant_id { get; set; } = 1;
+        public long tenant_id { get; set; } = 1;
 
 
         #endregion
